RpcApi: Change max_power parameter type from int to float in setDCCharging methods

+++ modules/RpcApi/rpc/RequestHandlerInterface.hpp
@@ -21,8 +21,8 @@ public:
     virtual types::json_rpc_api::ErrorResObj setACCharging(const int32_t evse_index, bool charging_allowed, bool max_current, std::optional<int> phase_count) = 0;
     virtual types::json_rpc_api::ErrorResObj setACChargingCurrent(const int32_t evse_index, float max_current) = 0;
     virtual types::json_rpc_api::ErrorResObj setACChargingPhaseCount(const int32_t evse_index, int phase_count) = 0;
-    virtual types::json_rpc_api::ErrorResObj setDCCharging(const int32_t evse_index, bool charging_allowed, int max_power) = 0;
-    virtual types::json_rpc_api::ErrorResObj setDCChargingPower(const int32_t evse_index, int max_power) = 0;
+    virtual types::json_rpc_api::ErrorResObj setDCCharging(const int32_t evse_index, bool charging_allowed, float max_power) = 0;
+    virtual types::json_rpc_api::ErrorResObj setDCChargingPower(const int32_t evse_index, float max_power) = 0;
     virtual types::json_rpc_api::ErrorResObj enableConnector(const int32_t evse_index, int connector_id, bool enable, int priority) = 0;
 };
